Supabase_Backend/Backend/requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
# dotenv==0.0.5
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.3
fqdn==1.5.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
json5==0.9.25
jsonpointer==2.4
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nest-asyncio==1.6.0
overrides==7.7.0
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
platformdirs==4.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
runserver==1.0.4
Send2Trash==1.8.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
terminado==0.18.1
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.26.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.2
wheel==0.43.0
supabase==2.4.5
psycopg2==2.9.9